test(analysis): align metrics tests with owner boundaries
